+++ src/D2NG/D2GS/Packet/PingPacket.cs
@@ -7,7 +7,7 @@ namespace D2NG.D2GS.Packet
         public PingPacket() :
             base(
                 BuildPacket(
-                    D2gs.PING,
+                    (byte)D2gs.PING,
                     BitConverter.GetBytes(Environment.TickCount),
                     BitConverter.GetBytes(0),
                     BitConverter.GetBytes(0)
